requirements.txt
pandas==1.4.3
Django==4.1.3
djangorestframework==3.14.0
